requirements.txt
langchain
langchain-community
langchain-groq
langgraph
pydantic
yfinance
pandas
pandas-ta
numpy
python-dotenv
